requirements.txt
pyautogen==0.2.32
python-dotenv==1.0.1
flask==3.0.3
pydantic==2.8.2
pandas==2.2.2